Backend/requirements.txt
flask
flask-cors
requests
google-generativeai
pymongo
mutagen